Handles character data updates, registration form takes in game_acct_id

+++ web/applications/artfweb/models/db.py
@@ -48,7 +48,7 @@ service = Service()
 plugins = PluginManager()
 
 auth.settings.extra_fields['auth_user'] = [
-    Field('game_acct_id', 'integer')
+    Field('game_acct_id', 'bigint')
 ]
 
 ## create all tables needed by auth if not custom tables
